01_ModelDeployment/requirements.txt
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1733250440834/work
blinker==1.9.0
click==8.3.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_comm_1753453984/work
debugpy @ file:///D:/bld/bld/rattler-build_debugpy_1758162072/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1740384970518/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1746947292760/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1756729339227/work
filelock==3.13.1
Flask==3.1.2
fsspec==2024.6.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
ipykernel @ file:///D:/bld/ipykernel_1754352877088/work
ipython @ file:///D:/bld/bld/rattler-build_ipython_1759151854/work
ipython_pygments_lexers @ file:///home/conda/feedstock_root/build_artifacts/ipython_pygments_lexers_1737123620466/work
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1733300866624/work
Jinja2==3.1.4
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1733440914442/work
jupyter_core @ file:///D:/bld/jupyter_core_1748333031907/work
MarkupSafe==2.1.5
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1733416936468/work
mpmath==1.3.0
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1733325553580/work
networkx==3.3
numpy==2.1.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
parso @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_parso_1755974222/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1733327343728/work
pillow==11.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1756227402/work
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1756321756983/work
psutil @ file:///D:/bld/psutil_1758169202226/work
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1733569405015/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pywin32 @ file:///D:/bld/bld/rattler-build_pywin32_1756487139/work
pyzmq @ file:///D:/bld/bld/rattler-build_pyzmq_1757387025/work
setuptools==78.1.1
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
stack_data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1733569443808/work
sympy==1.13.3
torch==2.8.0+cu129
torchvision==0.23.0+cu129
tornado @ file:///D:/bld/tornado_1756854854126/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1733367359838/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1758622279606/work
Werkzeug==3.1.3
wheel==0.45.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
